Nice configuration system.
How does it work?

conf.erl defines a behavior, which forces any implementing module to
define a default_conf/0 function. The function must yield a list of
parameters which are tought to be "element(3, {M,F,A})".

Workers spawned by the supervisors are supposed to have the behavior(conf)
directive.