Bugfix + Enhancement
This patch modifies the src/utils module in order to add a callback. The
callback allows to tweak the behavior of both gen_keeper and gen_peer in a
user-independent way.

For gen_peer this is used to initialize the random generator;
For gen_keeper this is used to do an early registration through
erlang:register/2.